fix makefile

+++ libft/ft_atoi_shlvl.c
@@ -6,17 +6,17 @@
 /*   By: mghalmi <mghalmi@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/07/24 15:00:26 by mghalmi           #+#    #+#             */
-/*   Updated: 2023/08/07 05:18:43 by mghalmi          ###   ########.fr       */
+/*   Updated: 2023/08/08 05:59:09 by mghalmi          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-unsigned int	ft_atoi_shlvl(char *str)
+unsigned long long	ft_atoi_shlvl(char *str)
 {
-	unsigned int	res;
-	unsigned int	i;
-	unsigned int	signe;
+	unsigned long long	res;
+	unsigned long long	i;
+	unsigned long long	signe;
 
 	res = 0;
 	i = 0;
